Add AI Task platform to OpenAI
